requirements.txt
nltk
langdetect
google-cloud-translate
tenacity
flask
flask_cors
openai
torch
pandas
transformers
python-dotenv